+++ kubeconfig.go
@@ -2,6 +2,9 @@ package kubeconfig
 
 import (
 	"context"
+	"encoding/base64"
+	"fmt"
+	yaml "gopkg.in/yaml.v2"
 
 	"github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/microerror"
@@ -62,6 +65,61 @@ func (k *KubeConfig) NewRESTConfigForApp(ctx context.Context, app v1alpha1.App) 
 	return restConfig, nil
 }
 
+func (k *KubeConfig) NewRESTConfigForKubeConfig(ctx context.Context, kubeConfig []byte) (*rest.Config, error) {
+	restConfig, err := clientcmd.RESTConfigFromKubeConfig(kubeConfig)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+	return restConfig, nil
+}
+
+func (k *KubeConfig) NewKubeConfigForRESTConfig(ctx context.Context, config *rest.Config, clusterName string) ([]byte, error) {
+	if clusterName == "" {
+		return nil, microerror.Maskf(executionError, "clusterName must not be empty")
+	} else if config == nil {
+		return nil, microerror.Maskf(executionError, "config must not be empty")
+	}
+
+	kubeConfig := KubeConfigValue{
+		APIVersion: "v1",
+		Kind:       "Config",
+		Clusters: []KubeconfigNamedCluster{
+			{
+				Name: clusterName,
+				Cluster: KubeconfigCluster{
+					Server:                   config.Host,
+					CertificateAuthorityData: base64.StdEncoding.EncodeToString(config.TLSClientConfig.CAData),
+				},
+			},
+		},
+		Contexts: []KubeconfigNamedContext{
+			{
+				Name: fmt.Sprintf("%s-context", clusterName),
+				Context: KubeconfigContext{
+					Cluster: clusterName,
+					User:    fmt.Sprintf("%s-user", clusterName),
+				},
+			},
+		},
+		Users: []KubeconfigUser{
+			{
+				Name: fmt.Sprintf("%s-user", clusterName),
+				User: KubeconfigUserKeyPair{
+					ClientCertificateData: base64.StdEncoding.EncodeToString(config.CertData),
+					ClientKeyData:         base64.StdEncoding.EncodeToString(config.KeyData),
+				},
+			},
+		},
+		CurrentContext: fmt.Sprintf("%s-context", clusterName),
+	}
+
+	bytes, err := yaml.Marshal(kubeConfig)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+	return bytes, nil
+}
+
 // getKubeConfigFromSecret returns KubeConfig bytes based on the specified secret information.
 func (k *KubeConfig) getKubeConfigFromSecret(ctx context.Context, secretName, secretNamespace string) ([]byte, error) {
 	secret, err := k.k8sClient.CoreV1().Secrets(secretNamespace).Get(secretName, metav1.GetOptions{})
@@ -77,4 +135,21 @@ func (k *KubeConfig) getKubeConfigFromSecret(ctx context.Context, secretName, se
 	} else {
 		return nil, microerror.Maskf(notFoundError, "Secret %#q in Namespace %#q does not have kubeConfig key in its data", secretName, secretNamespace)
 	}
+}
+
+func marshal(config *KubeConfigValue) ([]byte, error) {
+	bytes, err := yaml.Marshal(config)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+	return bytes, nil
+}
+
+func unmarshal(bytes []byte) (*KubeConfigValue, error) {
+	var kubeConfig KubeConfigValue
+	err := yaml.Unmarshal(bytes, &kubeConfig)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+	return &kubeConfig, nil
 }
